feat: lazy initial option values
All Options, except for the Counter, now accept lazy initial values. If the initial value is a function, the function will be called and the returned value resolved, if it's a promise. The function will not be called if the value has been provided by other means.

BREAKING CHANGE: Cli.prototype.process() now returns a Promise